♻️ Refactor: pages 폴더 구분 및 CartDetails, Productlist 컴포넌트로 분리
