+++ src/basic/__tests__/basic.test.tsx
@@ -8,8 +8,8 @@ import {
   screen,
   within,
 } from '@testing-library/react'
-import { CartPage } from '../../refactoring/components/CartPage'
-import { AdminPage } from '../../refactoring/components/AdminPage'
+import { CartPage } from '../../refactoring/pages/CartPage'
+import { AdminPage } from '../../refactoring/pages/AdminPage'
 import { CartItem, Coupon, Product } from '../../types'
 import { useCart, useCoupons, useProducts } from '../../refactoring/hooks'
 import * as cartUtils from '../../refactoring/models/cart'
